+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added id mapped dataset class
 - Added canonical name property 
 - Added dask support
+- Added dataset names
 
 ## [0.1.1] - 2023-03-27